also move PairwiseCBMAEstimator
